# Conflicts:
#	.idea/modules.xml
#	app/src/main/java/amsi/dei/estg/ipleiria/pt/faturasol/Classes/SingletonF_OL.java
#	app/src/main/java/amsi/dei/estg/ipleiria/pt/faturasol/Registar.java

+++ app/src/main/java/amsi/dei/estg/ipleiria/pt/faturasol/Classes/FaturaDBHelper.java
@@ -446,7 +446,7 @@ import java.util.List;
         Cursor cursor = this.database.rawQuery(sql,null);
         if(cursor.moveToFirst()){
             do{
-                Linha_Fatura auxLinhaFatura = new Linha_Fatura(cursor.getInt(0),
+                Linha_Fatura auxLinhaFatura = new Linha_Fatura(cursor.getLong(0),
                         cursor.getFloat(1),
                         cursor.getString(2),
                         cursor.getString( 3),
